Remove _NS and .pot.
