--- l10n_es_tgss/models/common.py
@@ -1,5 +1,0 @@
-# -*- coding: utf-8 -*-
-# © 2015 Grupo ESOC Ingeniería de servicios, S.L.
-# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
-
-_NS = __name__.split(".")[-3]
